reorganized entities into player and enemy folders
