+++ core/src/main/java/io/github/lobster/basicshootergame/entities/enemy/EnemyEntity.java
@@ -1,9 +1,11 @@
-package io.github.lobster.basicshootergame.entities;
+package io.github.lobster.basicshootergame.entities.enemy;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Vector2;
+
+import io.github.lobster.basicshootergame.entities.player.PlayerEntity;
 
 public abstract class EnemyEntity {
 	protected Vector2 position;
